drop use of libudev in favour of udevadm output parsing
Using libudev makes the packaging more complicated (now we need
to ship libudev in the snap) and is also not really needed because
we do not use any of the dynamic niceness of libudev. We just use
it to detect removable devices which we can equally well do with
the output of udevadm.

So instead of using a cool and small cgo based libgudev wrapper
we go back to good old text parsing.